feat: attach or create sessions
Changed the function to first check for existing sessions, and attach if
they do. This is because I generally use the `config` and `work`
sessions daily, however I would not like to have an overload of sessions,
especially if I'm not using them. This kind of forces me to finish my
work in the existing session before I get to `config` and `work`.