feat: git branch without traversing to file
In the last commit, the git branch did not show when I was editing a
file outside the git repository. But now, due to expanding the "%:h"
parameter (the full path and head of file), git now checks the directory
of the file instead of $PWD, which accurately shows the branch of the
file.