feat: draft function to stage, commit and push
This is a draft function to make using git easier.
The while loop is faulty because the input is taken outside it, however
the rest of it should work fine.

"gcp" can automatically add all files, or only add specific files.
Then, it can either let the user use a text editor, or directly type the
commit message in the terminal.
Finally, it will ask for a specific remote and branch, which are origin
and main by default.