refactor: re-write portion to check "$PATH"
Instead of printing out the entire "$PATH", I now simply check if
"$HOME/bin" exists in the "$PATH". This shortens the output (after my
experience with WSL2). However, the problem with this approach is that
it spawns two recursive subshells (which, common sense dictates, is a
terrible idea). I will attempt to fix this issue.