feat: script to enter latest directory/make new directory
This script is created specifically for my journal, which is sorted
according to the date. This script allows me to create a new file inside
a directory created with the current date at time of creation. It also
lets me enter the last directory if I want to do that (autocompletion
hardly works when so many characters of different directories are the
same)