!refactor: major change to "tld.sh" workflow (mtd.sh)
The commit is a bit deceptive, but here are the changes:

Rename "tld.sh" to "mtd.sh".
Only keep the logic for creating a directory and a file inside that in "mtd.sh".
Create an alias "tld" to move to the directory named with the current date. Only works if such a directory exists (like after running "mtd.sh").